requirements.txt
anyio==3.6.2
APScheduler==3.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
dnspython==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
ptb==0.0.3
Pygments==2.0.2
pymongo==4.3.3
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tzdata==2022.7
tzlocal==4.2
